Load game logic
Changed Position in GamePosition from double to int.
Added logic to load an existing game from the database using its name.

+++ Source/GameEngine/DbModels/GamePosition.cs
@@ -17,6 +17,6 @@ namespace GameEngine.Models
         public User User { get; set; }
 
         [Required]
-        public double Position { get; set; } = 0;
+        public int Position { get; set; } = 0;
     }
 }
